restructure web api controllers for unique namespaces to avoid possible naming conflicts
