+++ GvtcGh.Web/Api/GHRestController.cs
@@ -9,7 +9,7 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Web.Http;
 
-namespace GvtcGh.Web.Controllers
+namespace GvtcGh.Web.Api
 {
     public class GHRestController : ApiController
     {
