+++ GvtcGh.Web/Api/RepositoryController.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Web.Http;
 
-namespace GvtcGh.Web.Controllers
+namespace GvtcGh.Web.Api
 {
     public class RepositoryController : ApiController
     {
