+++ GvtcGh.Web/Controllers/GitHubController.cs
@@ -5,6 +5,7 @@ using System.Dynamic;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
+using GvtcGh.Web.Api;
 using GvtcGh.Web.Models;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
